backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
openpyxl>=3.1
PyMySQL>=1.1
gunicorn>=22.0
